requirements.txt
coinbase-advanced-py==1.8.1
python-telegram-bot==21.7

